+++ source/scene/Text.h
@@ -43,7 +43,7 @@ namespace scene
         const renderer::RenderJobPtr&   getRenderJob() const;
 
         void                            init()                          override;
-        void                            update(f64 time)                override;
+        void                            update(s32 time)                override;
         void                            render()                        override;
 
     protected:
